requirements.txt
pytest
pyyaml<5.1
